+++ R/BTm.setup.R
@@ -8,7 +8,9 @@ BTm.setup <- function(player1, player2 = NULL, outcome = 1, formula = NULL,
                                    deparse(substitute(player2)))
         if (!length(keep)) keep <- FALSE
         ## save row names for checking against index variables (in Diff)
-        data <- lapply(data, as.data.frame)
+        data[keep] <- lapply(data[keep], as.data.frame)
+        if (!all(vapply(data[!keep], inherits, logical(1), "data.frame")))
+            stop("unnamed elements of 'data' should be data.frames")
         nm <- lapply(data, rownames)
         data <- c(data[keep], unlist(unname(data[!keep]), recursive = FALSE))
         if (any(dup <- duplicated(names(data))))
@@ -16,35 +18,32 @@ BTm.setup <- function(player1, player2 = NULL, outcome = 1, formula = NULL,
                     paste(names(data)[dup], collapse = " "))
     }
     ## (will take first occurence of replicated names)
-    withIfNecessary <- function(x, data, as.data.frame = TRUE) {
+    withIfNecessary <- function(x, formula, data = NULL, as.data.frame = TRUE) {
         if (as.data.frame)
             expr <- substitute(data.frame(x), list(x = x))
         else expr <- x
-        if (!is.null(data))
-            with(data, eval(expr))
-        else eval(expr)
+        eval(expr, data, enclos = environment(formula))
     }
-    player1 <- withIfNecessary(substitute(player1), data)
+    player1 <- withIfNecessary(substitute(player1), formula, data)
     if (ncol(player1) == 1) colnames(player1) <- id
     if (ncol(player1) == 0) player1 <- NULL
-    player2 <- withIfNecessary(substitute(player2), data)
+    player2 <- withIfNecessary(substitute(player2), formula, data)
     if (ncol(player2) == 1) colnames(player2) <- id
     if (ncol(player2) == 0) player2 <- NULL
-    Y <- withIfNecessary(substitute(outcome), c(player1, player2, data),
-                         as.data.frame = FALSE)
-    weights <- withIfNecessary(substitute(weights), data, FALSE)
-    subset1 <- withIfNecessary(substitute(subset),
+    Y <- withIfNecessary(substitute(outcome), formula, 
+                         c(player1, player2, data), as.data.frame = FALSE)
+    weights <- withIfNecessary(substitute(weights), formula, data, FALSE)
+    subset1 <- withIfNecessary(substitute(subset), formula, 
                                c(player1 = list(player1),
                                  player2 = list(player2), player1, data), FALSE)
-    subset2 <- withIfNecessary(substitute(subset),
+    subset2 <- withIfNecessary(substitute(subset), formula, 
                                c(player1 = list(player1),
                                  player2 = list(player2), player2, data), FALSE)
     if (is.logical(subset1)) subset <- subset1 | subset2
     else subset <- c(subset1, subset2)
-    if (is.null(formula)) formula <- reformulate(id)
     diffModel <- Diff(player1, player2, formula, id, data, separate.ability,
                       refcat, contrasts, nm)
-    offset <- withIfNecessary(substitute(offset), data, FALSE) #contest level
+    offset <- withIfNecessary(substitute(offset), formula, data, FALSE) #contest level
     if (!is.null(offset)) {
         if (is.null(diffModel$offset))  diffModel$offset <- offset
         else diffModel$offset <- diffModel$offset + offset