+++ R/model.matrix.BTm.R
@@ -1,5 +1,0 @@
-#' @importFrom stats model.frame
-#' @export
-model.matrix.BTm <- function(object, ...){
-    model.frame(object)$X
-}